requirements.txt
h5py==2.10.0
matplotlib==3.3.3
networkx==2.5
numpy==1.19.2
pandas==1.1.5
PyYAML==6.0
scipy==1.5.2
tensorboardX==2.4.1
torch==1.6.0
torch_geometric==1.6.3
torch_sparse==0.6.8
tqdm==4.55.1
